[fix] id 자료형 long으로 수정

+++ backend/learnacademy/src/main/kotlin/com/ssafy/learnacademy/vo/Academy.kt
@@ -10,7 +10,7 @@ import javax.persistence.*
 class Academy (
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
-        var academyId: Int? = null,
+        var academyId: Long? = null,
 
         @Column(nullable = false)
         var name: String? = null,
